+++ server/src/main/java/at/ac/tgm/config/AdLdapConfig.java
@@ -1,11 +1,6 @@
 package at.ac.tgm.config;
 
 import at.ac.tgm.ad.Roles;
-import com.fasterxml.jackson.core.JsonGenerator;
-import com.fasterxml.jackson.databind.ObjectMapper;
-import com.fasterxml.jackson.databind.SerializerProvider;
-import com.fasterxml.jackson.databind.module.SimpleModule;
-import com.fasterxml.jackson.databind.ser.std.StdSerializer;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
@@ -20,10 +15,6 @@ import org.springframework.security.ldap.authentication.ad.ActiveDirectoryLdapAu
 import org.springframework.security.ldap.userdetails.LdapUserDetailsMapper;
 import org.springframework.security.ldap.userdetails.UserDetailsContextMapper;
 
-import javax.naming.Name;
-import java.io.IOException;
-import java.time.LocalDateTime;
-import java.time.format.DateTimeFormatter;
 import java.util.*;
 
 @Configuration
@@ -69,45 +60,5 @@ public class AdLdapConfig {
     @Bean
     public AuthenticationManager authenticationManager(ActiveDirectoryLdapAuthenticationProvider adProvider) {
         return new ProviderManager(Collections.singletonList(adProvider));
-    }
-    
-    @Bean
-    public ObjectMapper registerObjectMapper() {
-        ObjectMapper mapper = new ObjectMapper();
-        SimpleModule module = new SimpleModule("MyObjectSerializer");
-        module.addSerializer(Name.class, new NameJsonSerializer());
-        module.addSerializer(LocalDateTime.class, new LocalDateTimeJsonSerializer());
-        mapper.registerModule(module);
-        return mapper;
-    }
-    
-    static class NameJsonSerializer extends StdSerializer<Name> {
-        public NameJsonSerializer() {
-            this(null);
-        }
-        
-        public NameJsonSerializer(Class<Name> t) {
-            super(t);
-        }
-        
-        @Override
-        public void serialize(Name value, JsonGenerator jgen, SerializerProvider provider) throws IOException {
-            jgen.writeString(value.toString());
-        }
-    }
-    
-    static class LocalDateTimeJsonSerializer extends StdSerializer<LocalDateTime> {
-        public LocalDateTimeJsonSerializer() {
-            this(null);
-        }
-        
-        public LocalDateTimeJsonSerializer(Class<LocalDateTime> t) {
-            super(t);
-        }
-        
-        @Override
-        public void serialize(LocalDateTime value, JsonGenerator jgen, SerializerProvider provider) throws IOException {
-            jgen.writeString(value.format(DateTimeFormatter.ofPattern("yyyy-MM-dd HH:mm:ss")));
-        }
     }
 }
